Rename secure -> config and changes its focus
