+++ cpp/Ice/config/Chatbot.cpp
@@ -6,9 +6,10 @@
 #include <sstream>
 
 using namespace std;
+using namespace ConfigServer;
 
 string
-GreeterServer::Chatbot::greet(string name, const Ice::Current&)
+Chatbot::greet(string name, const Ice::Current&)
 {
     cout << "Dispatching greet request { name = '" << name << "' }" << endl;
