+++ cpp/Ice/config/config.client
@@ -1,0 +1,38 @@
+#
+# Application-specific configuration
+#
+
+# The address of the target object. `default` is replaced by the value of Ice.Default.Protocol.
+Greeter.Proxy=greeter:default -h localhost -p 4061
+
+#
+# Ice configuration
+#
+
+# The transport protocol to use for Ice communications when `default` is specified.
+# You can set it to tcp, ssl, ws, or wss. Leaving it unset is equivalent to setting it to tcp.
+# Make sure you use the same value in config.client and config.server.
+Ice.Default.Protocol=ssl
+
+#
+# IceSSL configuration
+#
+
+# The directory where the application finds certificates and other files.
+IceSSL.DefaultDir=../../../certs
+
+# The certificate authority file.
+IceSSL.CAs=cacert.pem
+
+# The client's certificate file.
+IceSSL.CertFile=client.p12
+
+# The name of the keychain in which to import the client's certificate (SecureTransport only).
+IceSSL.Keychain=client.keychain
+
+# TODO: switch to password-less certificates
+IceSSL.Password=password
+IceSSL.KeychainPassword=password
+
+# Turn on security logging/tracing.
+# IceSSL.Trace.Security=1
