--- cpp/Ice/secure/README.md
@@ -1,39 +1,0 @@
-# Secure
-
-This demo shows how to configure client and server applications to communicate over
-secure connections.
-
-To build the demo run:
-
-```shell
-cmake -B build
-cmake --build build --config Release
-```
-
-To run the demo, first start the server:
-
-**Linux/macOS:**
-
-```shell
-./build/server
-```
-
-**Windows:**
-
-```shell
-build\Release\server
-```
-
-In a separate window, start the client:
-
-**Linux/macOS:**
-
-```shell
-./build/client
-```
-
-**Windows:**
-
-```shell
-build\Release\client
-```
