--- cpp/Ice/secure/config.client
@@ -1,19 +1,0 @@
-# The directory where the application finds certificates and other files.
-IceSSL.DefaultDir=../../../certs
-
-# The certificate authority file.
-IceSSL.CAs=cacert.pem
-
-# The client's certificate file.
-IceSSL.CertFile=client.p12
-
-# The name of the keychain in which to import the client's certificate (SecureTransport only).
-IceSSL.Keychain=client.keychain
-
-# TODO: switch to password-less certificates
-IceSSL.Password=password
-IceSSL.KeychainPassword=password
-
-# Turn on security and network tracing
-IceSSL.Trace.Security=1
-Ice.Trace.Network=1
