--- cpp/Ice/secure/config.server
@@ -1,19 +1,0 @@
-# The directory where the application finds certificates and other files.
-IceSSL.DefaultDir=../../../certs
-
-# The certificate authority file.
-IceSSL.CAs=cacert.pem
-
-# The server's certificate file.
-IceSSL.CertFile=server.p12
-
-# The name of the keychain in which to import the server's certificate (SecureTransport only).
-IceSSL.Keychain=server.keychain
-
-# TODO: switch to password-less certificates
-IceSSL.Password=password
-IceSSL.KeychainPassword=password
-
-# Turn on security and network tracing
-IceSSL.Trace.Security=1
-Ice.Trace.Network=1
